+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Waiting for Release
 - Enhance UI for Go to render call
 - Set "Alt + R" by default shortcut for Go to View render call
 - Send plugin exceptions from IDE Error Report Dialog
+- Applied migration count
 
 0.9.50 May 17, 2018
 -------------------